Split out settings elements into another file
